moved most of the dropbox operaitons
