+++ src/index.js
@@ -1,5 +1,6 @@
 import { AuthenticaionService } from "./services/auth.service";
 import { ExtractionService } from "./services/extraction.service";
+import { DropboxService } from "./services/dropbox.service";
 
 const shared = {
   counter: undefined,
@@ -23,6 +24,7 @@ const shared = {
 }
 
 const extractionService = new ExtractionService(shared);
+const dropboxService = new DropboxService(shared);
 
 //! Dropbox credentials
 
@@ -253,7 +255,7 @@ async function filesSearch(studentNumber, taskName) {
                 dlIcon.classList.add("rotate-center");
                 dlIcon.src = chrome.runtime.getURL("images/loader.png");
                 //! Download selected Folder or file
-                downloadFolder(item.metadata.metadata, dlIcon); //folder
+                dropboxService.downloadFolder(item.metadata.metadata, dlIcon, dbx); //folder
               });
 
               dlIconContainer.appendChild(dlIcon);
@@ -297,62 +299,6 @@ async function filesSearch(studentNumber, taskName) {
       return;
     });
 }
-
-//Download the Folder
-function downloadFolder(dir, dlIcon) {
-  console.log(`%c  creating Download Folder path`, "color: red");
-  let pdfExists = false;
-  if (dir.path_display.endsWith(".pdf")) {
-    pdfExists = true;
-  }
-
-  //download the found files parent directory
-  if (pdfExists) {
-    const lastIndex = dir.path_display.lastIndexOf("/");
-    const folderPath = dir.path_display.substring(0, lastIndex);
-    downloadFileBob(folderPath, dlIcon);
-  } else {
-    //download the  directory
-    downloadFileBob(dir.path_display, dlIcon);
-  }
-}
-
-//download the selected fine in zip format
-async function downloadFileBob(path, dlIcon) {
-  await dbx
-    .filesDownloadZip({ path: path })
-    .then(function (response) {
-      //const fileName = response.result.metadata.name
-      const blob = new Blob([response.result.fileBlob], {
-        type: "application/zip",
-      });
-      getDLLink(blob, path.substring(path.lastIndexOf("/") + 1));
-      //displayFiles(response.result.fileBlob);
-    })
-    .catch(function (error) {
-      console.log(error);
-      alert(
-        "Error downloading file. Download folder could be containing more than 100 files"
-      );
-    });
-
-  dlIcon.classList.remove("rotate-center");
-  dlIcon.src = chrome.runtime.getURL("images/dlFOlder.png");
-}
-
-//Add href download link for folder to button
-function getDLLink(blob, name) {
-  const url = window.URL.createObjectURL(blob);
-  const link = document.createElement("a");
-  link.href = url;
-  link.download = studentName + "_" + name;
-  document.body.appendChild(link);
-  document.body.appendChild(link);
-  link.click();
-  document.body.removeChild(link);
-  //loadingImage.style.display = "none";
-}
-
 //Extracts the words that matches the task name and only highlight those words.
 function highlightTaskName(taskName) {
